fix: 프로젝트 전체 검증 — 버그 수정 및 품질 개선
Critical:
- HttpServer: SocketException catch 추가 (포트 충돌 시 서버 중단 방지)
- ExecuteCsharp: return 감지 regex 수정 (\breturn[\s;] → \breturn\b)
- ExecuteCsharp: 컴파일러 명령줄 길이 초과 수정 (응답 파일 .rsp 사용)

Moderate:
- ReadConsole: reflection 초기화 실패 시 누락 멤버 특정
- ManageProfiler: CollectFlat calls 캐스트 int→long

Improvements:
- ManageEditor: SerializedObject using 블록 추가
- StringCaseUtility: 연속 대문자 처리 (HTTPClient→http_client)
- ReadConsole: logEntryType 정적 캐시

+++ unity-connector/Editor/Tools/ManageProfiler.cs
@@ -252,7 +252,7 @@ namespace UnityCliConnector.Tools
                 var name = frameData.GetItemName(childId);
                 var totalMs = frameData.GetItemColumnDataAsFloat(childId, HierarchyFrameDataView.columnTotalTime);
                 var selfMs = frameData.GetItemColumnDataAsFloat(childId, HierarchyFrameDataView.columnSelfTime);
-                var calls = (int)frameData.GetItemColumnDataAsFloat(childId, HierarchyFrameDataView.columnCalls);
+                var calls = (long)frameData.GetItemColumnDataAsFloat(childId, HierarchyFrameDataView.columnCalls);
 
                 if (acc.TryGetValue(name, out var existing))
                     acc[name] = (existing.totalMs + totalMs, existing.selfMs + selfMs, existing.calls + calls, existing.count + 1);
